+++ NEWS.md
@@ -1,3 +1,7 @@
+# gasanalysr 0.0.2.9000 2022-09-30
+## Added
+* functions to help processing files exported from LumasoftSoftGas 7880 photo acoustic gasmonitor 
+
 # gasanalysr 0.0.1.9000 2022-05-06
 ## Added
 * a function to convert gas concentration between mg/m3 and ppm